samples/durable-task-sdks/python/opentelemetry-tracing/requirements.txt
durabletask
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-grpc
